remove redundant imports and move default objects to point_defaults
