+++ react/components/Filter/Edit_spatial_filter.js
@@ -5,11 +5,9 @@ import {
   FeatureGroup
 } from "react-leaflet";
 import { EditControl } from "react-leaflet-draw";
-//import 'leaflet/dist/leaflet.css';
 import 'leaflet-draw/dist/leaflet.draw.css';
-//import 'leaflet-draw';
 
-export const RenderSpatialFilter = forwardRef(function RenderSpatialFilter ({polyFilterCoords, setPolyFilterCoords}, ref) {
+export const RenderSpatialFilter = forwardRef(function RenderSpatialFilter ({ setPolyFilterCoords}, ref) {
   const featureGroupRef = useRef(null);
   // Get the FeatureGroup reference
   function clearShapes(){
