+++ react/components/Filter/dateFilter.js
@@ -3,10 +3,10 @@ import dayjs from 'dayjs';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import CalendarMonthOutlinedIcon from '@mui/icons-material/CalendarMonthOutlined';
-import { useEffect, useState } from 'react';
-import { IconButton, TextField, Box, Typography, Paper, Button, Stack, CardContent, Card, CardActions } from "@mui/material";
+import {  useState } from 'react';
+import {  Box,  Button,  CardContent, Card, CardActions } from "@mui/material";
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
-import {Slider, Tooltip} from '@mui/material';
+import {Slider} from '@mui/material';
 import { smallScreenIconButton } from './filter_utils';
