+++ react/components/Filter/filter.js
@@ -1,5 +1,5 @@
-import { IconButton, TextField, Box, Typography, Paper, Button, SpeedDial, SpeedDialIcon, SpeedDialAction, Tooltip } from "@mui/material";
-import { useEffect, useState, Fragment } from 'react';
+import { IconButton,  Box, Typography, Paper, Button, SpeedDial,  SpeedDialAction,  } from "@mui/material";
+import {  useState } from 'react';
 import Stack from '@mui/material/Stack';
 
 //import Chip from '@mui/material/Chip';
