+++ react/components/Filter/filter_utils.js
@@ -1,4 +1,4 @@
-import { IconButton, TextField, Box, Typography, Paper, Button, SpeedDial,  SpeedDialIcon, SpeedDialAction, Tooltip } from "@mui/material";
+import { IconButton, Tooltip } from "@mui/material";
 
 export function smallScreenIconButton(toolTipTitle, handleIconClick, buttonStyle, Icon){
   return (
