+++ react/components/historical_storm/render_search_result.js
@@ -1,6 +1,5 @@
-import { handleClick } from "./historical_storm_utils";
 import { Button } from "@mui/material";
-import { IconButton, TextField, Box, Typography, Paper } from "@mui/material";
+import {  Box, Typography, Paper } from "@mui/material";
 import { Stack } from "@mui/system";
 import { handleStormButtonClick } from "./historical_storm_utils";
