+++ react/components/loading_screen.js
@@ -1,5 +1,4 @@
 import React from "react";
-import CircularProgress from '@mui/material/CircularProgress';
 import LinearProgress from '@mui/material/LinearProgress';
 import Box from '@mui/material/Box';
