+++ react/components/render_storm_search.js
@@ -1,4 +1,4 @@
-import { IconButton, TextField, Box, Typography, Paper, Button } from "@mui/material";
+import { IconButton,  Box,  Button } from "@mui/material";
 import Search from "@mui/icons-material/Search";
 //import { StormSearchQuery } from "./search_storm_header";
 import { handleFormSubmit } from "./historical_storm/historical_storm_utils";
