+++ react/components/search_storm_in_header.js
@@ -1,7 +1,6 @@
-import { handleSearch, handleFormSubmit } from "./historical_storm/historical_storm_list";
-import { IconButton, TextField, Box, Typography, Paper } from "@mui/material";
+import {  handleFormSubmit } from "./historical_storm/historical_storm_list";
+import { IconButton,  Box, Typography, Paper } from "@mui/material";
 import Search from "@mui/icons-material/Search";
-import { useEffect, useState } from 'react';
 import { Stack } from "@mui/system";
 import {handleStormNameClick} from '../pages/about_page'
 import { useRouter } from 'next/router';
