+++ react/components/station_dashboard/station_dashboard.js
@@ -1,18 +1,10 @@
-import React, { useState, useRef } from"react";
+import React from"react";
 import { FaWindowClose } from "react-icons/fa";
-import { empty_station_obj } from "../layout";
 import { useMediaQuery, Box, useTheme } from "@mui/material";
-//import { empty_station_obj } from "../layout";
-//import { useMediaQuery, Box } from "@mui/material";
 import StationDataLayout from "./station_layout_small_screen";
 import BasicTabs from "./tabs";
-//import { RecentStationData, getMatchedStation } from "../utils/station_data_format_util";
-//import BasicTabs from "./tabs";
-import { RecentStationData, getMatchedStation, getStationDataText, } from "../utils/station_data_format_util";
-//import BasicTabs from "./tabs";
+import {  getMatchedStation, getStationDataText, } from "../utils/station_data_format_util";
 import { fetch_value } from "@/lib/storm_utils";
-import { RenderSmallDashboard } from "../Dashboard/Mobile_Dashboard/dashboard_small";
-import { RowingSharp } from "@mui/icons-material";
 
 
 /**
