+++ react/components/station_marker.js
@@ -1,14 +1,8 @@
 import React, { useState, useMemo } from "react";
-import station_names from "../data/station/names.json"
-import { RecentStationData } from './utils/station_data_format_util'
-import RenderChart from './station_graph.js'
 import { flip_coords } from "@/lib/storm_utils";
-import { empty_station_obj } from "./layout"
-
 import { Marker, Tooltip, Popup, Icon } from "react-leaflet";
-
 import { getDisplayName } from "./utils/station_data_format_util";
-//import styles from './station_marker.module.css'
+
