+++ react/components/storm_dashboard/storm_details.js
@@ -1,6 +1,6 @@
 import { parseISO, format } from 'date-fns';
 import { fetch_value } from "@/lib/storm_utils";
-import React, {useState} from "react";
+import React from "react";
 import StormType from '../Storm_popup/Storm_type';
 import StormCategory from '../Storm_popup/storm_category';
 import StormPressure from '../Storm_popup/storm_pressure';
